backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# CORS
python-multipart>=0.0.17

# AI / LLM
google-genai

# YouTube Data API
google-api-python-client>=2.150.0

# MIDI Processing
mido>=1.3.3

# HTTP Client
httpx>=0.28.0

# Environment
python-dotenv>=1.0.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Type Hints
pydantic>=2.10.0
